Made RSPlayer a singleton since there can only be one player loaded at a time.
Added the ability to search for players by user name

Added checks to make sure the player is valid

Fixed up more of quest loader so that the quests and skills are added to a list for easy access. Still need to remove the GUI calls and move the printing somewhere else.

GUI formatting fixes